+++ test/ext/string_test.exs
@@ -1,0 +1,5 @@
+defmodule Ext.StringTest do
+  use ExUnit.Case, async: true
+
+  doctest Ext.String
+end